# server/src/main/java/cn/camera/safe/api/ErrorCode.java
package cn.camera.safe.api;

public enum ErrorCode {
    INVALID_REQUEST,
    INVALID_COORDINATE,
    UNSUPPORTED_COORDINATE_SYSTEM,
    OUTSIDE_ROUTING_BOUNDS,
    START_IN_RESTRICTED_AREA,
    END_IN_RESTRICTED_AREA,
    NO_COMPLIANT_ROUTE,
    ROUTE_SEARCH_TIMEOUT,
    ROUTE_SEARCH_RESOURCE_LIMIT,
    SIXTH_RING_TOPOLOGY_NOT_READY,
    SIXTH_RING_BOUNDARY_AMBIGUOUS,
    REFERENCE_ROUTE_FAILED,
    ROUTE_CONFLICT_DETECTED,
    ROUTING_NOT_READY,
    CAMERA_SNAPSHOT_NOT_READY,
    REFRESH_ALREADY_RUNNING,
    CAMERA_UPDATE_ALREADY_RUNNING,
    CAMERA_UPDATE_FAILED,
    INTERNAL_ERROR
}